20200416ABC

+++ ABC135A.py
@@ -3,4 +3,4 @@ if (b-a)%2 == 1:
     print('IMPOSSIBLE')
 
 else:
-    print((b-a)/2 + a)
+    print(int((b-a)/2 + a))
